requirements.txt
brother_ql==0.9.4
Keras==2.4.3
dlib==19.22.0
scipy==1.5.4
numpy==1.19.5
pandas==1.1.5
tensorflow==2.4.1
matplotlib==3.3.3
torchvision==0.7.0
pyrealsense2==2.41.0.2666
pyzbar==0.1.8
smbus==1.1.post2
gdown==3.12.2
libusb1==1.9.1
opencv_python==4.5.1.48
mtcnn==0.1.0
requests==2.25.1
tqdm==4.56.0
torch==1.6.0
Flask==1.1.2
eventlet==0.31.0
flask_socketio==5.0.3
GitPython==3.1.17
lightgbm==3.2.1
Pillow==8.2.0
pyserial==3.5
pyusb==1.1.1
PyYAML==5.4.1
seaborn==0.11.1
thop==0.0.31-2005241907
